Refactoring SubscriptionService
