+++ DeviceBridge/Services/IDataSubscriptionService.cs
@@ -7,22 +7,12 @@ using NLog;
 
 namespace DeviceBridge.Services
 {
-    public interface ISubscriptionService
+    public interface IDataSubscriptionService
     {
-        Task<DeviceSubscription> CreateOrUpdateConnectionStatusSubscription(Logger logger, string deviceId, string callbackUrl, CancellationToken cancellationToken);
-
         Task<DeviceSubscriptionWithStatus> CreateOrUpdateDataSubscription(Logger logger, string deviceId, DeviceSubscriptionType subscriptionType, string callbackUrl, CancellationToken cancellationToken);
-
-        Task DeleteConnectionStatusSubscription(Logger logger, string deviceId, CancellationToken cancellationToken);
 
         Task DeleteDataSubscription(Logger logger, string deviceId, DeviceSubscriptionType subscriptionType, CancellationToken cancellationToken);
 
-        Task<DeviceSubscription> GetConnectionStatusSubscription(Logger logger, string deviceId, CancellationToken cancellationToken);
-
         Task<DeviceSubscriptionWithStatus> GetDataSubscription(Logger logger, string deviceId, DeviceSubscriptionType subscriptionType, CancellationToken cancellationToken);
-
-        Task StartDataSubscriptionsInitializationAsync();
-
-        Task SynchronizeDeviceDbAndEngineDataSubscriptionsAsync(string deviceId, bool useInitializationList = false);
     }
 }
