+++ DeviceBridge/Startup.cs
@@ -59,8 +59,8 @@ namespace DeviceBridge
             var customRampupBatchSize = Environment.GetEnvironmentVariable("DEVICE_RAMPUP_BATCH_SIZE");
             var customRampupBatchIntervalMs = Environment.GetEnvironmentVariable("DEVICE_RAMPUP_BATCH_INTERVAL_MS");
             uint maxPoolSize = (customMaxPoolSize != null && customMaxPoolSize != string.Empty) ? Convert.ToUInt32(customMaxPoolSize, 10) : ConnectionManager.DeafultMaxPoolSize;
-            uint rampupBatchSize = (customRampupBatchSize != null && customRampupBatchSize != string.Empty) ? Convert.ToUInt32(customRampupBatchSize, 10) : SubscriptionService.DefaultRampupBatchSize;
-            uint rampupBatchIntervalMs = (customRampupBatchIntervalMs != null && customRampupBatchIntervalMs != string.Empty) ? Convert.ToUInt32(customRampupBatchIntervalMs, 10) : SubscriptionService.DefaultRampupBatchIntervalMs;
+            uint rampupBatchSize = (customRampupBatchSize != null && customRampupBatchSize != string.Empty) ? Convert.ToUInt32(customRampupBatchSize, 10) : SubscriptionScheduler.DefaultRampupBatchSize;
+            uint rampupBatchIntervalMs = (customRampupBatchIntervalMs != null && customRampupBatchIntervalMs != string.Empty) ? Convert.ToUInt32(customRampupBatchIntervalMs, 10) : SubscriptionScheduler.DefaultRampupBatchIntervalMs;
 
             _logger.SetProperty("idScope", idScope);
             _logger.SetProperty("cv", Guid.NewGuid()); // CV for all background operations
@@ -73,12 +73,16 @@ namespace DeviceBridge
             services.AddSingleton<IEncryptionService, EncryptionService>();
             services.AddSingleton<IStorageProvider>(provider => new StorageProvider(sqlConnectionString, provider.GetRequiredService<IEncryptionService>()));
             services.AddSingleton<IConnectionManager>(provider => new ConnectionManager(provider.GetRequiredService<Logger>(), idScope, sasKey, maxPoolSize, provider.GetRequiredService<IStorageProvider>()));
-            services.AddSingleton<ISubscriptionService>(provider => new SubscriptionService(provider.GetRequiredService<Logger>(), provider.GetRequiredService<IConnectionManager>(), provider.GetRequiredService<IStorageProvider>(), provider.GetRequiredService<IHttpClientFactory>(), rampupBatchSize, rampupBatchIntervalMs));
+            services.AddSingleton<ISubscriptionCallbackFactory, SubscriptionCallbackFactory>();
+            services.AddSingleton<IConnectionStatusSubscriptionService, ConnectionStatusSubscriptionService>();
+            services.AddSingleton<IDataSubscriptionService, DataSubscriptionService>();
+            services.AddSingleton<ISubscriptionScheduler>(provider => new SubscriptionScheduler(provider.GetRequiredService<Logger>(), provider.GetRequiredService<IConnectionManager>(), provider.GetRequiredService<IStorageProvider>(), provider.GetRequiredService<ISubscriptionCallbackFactory>(), rampupBatchSize, rampupBatchIntervalMs));
             services.AddSingleton<IBridgeService, BridgeService>();
             services.AddHttpClient("RetryClient").AddPolicyHandler(GetRetryPolicy(_logger));
 
             services.AddHostedService<ExpiredConnectionCleanupHostedService>();
             services.AddHostedService<SubscriptionStartupHostedService>();
+            services.AddHostedService<SubscriptionSchedulerHostedService>();
             services.AddHostedService<HubCacheGcHostedService>();
 
             services.AddAuthentication(o =>
